Pull request #1871: Provision: EFR32MG26 support.
Merge in WMN_TOOLS/matter from feature/provision_mg26 to silabs_slc_1.3

Squashed commit of the following:

commit aa0f5190ba13084bde2e968f393735b779029f22
Merge: afff7499fa e79dcfbf2f
Author: Sergei Lissianoi <selissia@silabs.com>
Date:   Wed May 22 17:48:54 2024 -0400

    Merge branch 'silabs_slc_1.3' into feature/provision_mg26

commit afff7499faf0a03065801ede3318e2148659272e
Author: Sergei Lissianoi <selissia@silabs.com>
Date:   Wed May 22 13:36:17 2024 -0400

    Add MG26 to the Jlink XML file, add loader bin

commit cee4ac82ec8c2ad8562ee2dd12d50133bc6f7c73
Author: Ricardo Casallas <77841255+rcasallas-silabs@users.noreply.github.com>
Date:   Tue May 21 16:48:53 2024 -0400

    Provision: EFR32MG26 support.

+++ provision/support.py
@@ -75,8 +75,8 @@ class Updater:
     SUPPORTED_PARTS = [
         PartConfig('efr32mg12', 'efr32', 'EFR32MG12P232F512GM68',     0x7f800),
         PartConfig('efr32mg24', 'efr32', 'EFR32MG24A010F1024IM48',    0xfe000),
+		PartConfig('efr32mg26', 'efr32', 'EFR32MG26B410F3200IM48',    0x31c000),
         PartConfig('si917',     'si917', 'brd4338a;wiseconnect3_sdk', 0x1fe000),
-
     ]
 
     def __init__(self, paths, args, part) -> None:
@@ -148,8 +148,9 @@ class Updater:
             patcher.patch(self.paths)
             # Linker file
             linkerfile_from = self.paths.support("patch/{}.ld".format(self.part.family))
-            linkerfile_into = "{}/autogen/linkerfile.ld".format(self.part_dir)
-            shutil.copyfile(linkerfile_from, linkerfile_into)
+            if os.path.exists(linkerfile_from):
+                linkerfile_into = "{}/autogen/linkerfile.ld".format(self.part_dir)
+                shutil.copyfile(linkerfile_from, linkerfile_into)
         # Stage the newly generated code
         _util.execute([ 'git', 'add', self.part_dir ])
